feat: Update maintenance and vehicle features
- Added distance unit string to maintenance strings.
- Changed maintenance mileage and next maintenance mileage types from double to int in the maintenance model.
- Updated maintenance form cubit to parse mileage as int.
- Enhanced maintenance form content with text input actions for better UX.
- Modified next maintenance mileage field to include text input action.
- Cleaned up vehicle list item widget by removing commented code.
- Adjusted maintenance card body and content to use int for remaining distance calculations.
- Added new vehicle strings for archived vehicles.
- Updated vehicle model to reorder enum values for consistency.
- Enhanced vehicle list page to display correct strings for archived vehicles.
- Improved vehicle form with updated icons and text input actions.
- Refactored shared form widgets to remove deprecated label widget and use a new text field label widget.
- Introduced authentication form fields constants for better maintainability.
- Created login and signup email forms with validation and submission logic.
- Added signup header and social buttons for improved signup experience.
- Implemented password text field with toggle visibility functionality.

+++ lib/features/maintenance/domain/model/maintenance_model.dart
@@ -27,10 +27,10 @@ class MaintenanceModel {
   final String? notes;
   final DateTime date;
   final DateTime? nextMaintenanceDate;
-  final double maintanceMileage;
+  final int maintanceMileage;
   final DistanceUnit distanceUnit;
   final bool receiveAlert;
-  final double? nextMaintenanceMileage;
+  final int? nextMaintenanceMileage;
   final DateTime? createdDate;
   final DateTime? updatedDate;
 
@@ -58,10 +58,10 @@ class MaintenanceModel {
     String? notes,
     DateTime? date,
     DateTime? nextMaintenanceDate,
-    double? maintanceMileage,
+    int? maintanceMileage,
     DistanceUnit? distanceUnit,
     bool? receiveAlert,
-    double? nextMaintenanceMileage,
+    int? nextMaintenanceMileage,
     String? userId,
     String? vehicleId,
     DateTime? createdDate,
