+++ lib/features/maintenance/presentation/widgets/item_card/maintenance_card_body.dart
@@ -14,7 +14,7 @@ class MaintenanceCardBody extends StatelessWidget {
   final double? progressPercent;
   final bool isUrgent;
   final int? daysUntilNext;
-  final double? Function(int?) getRemainingDistance;
+  final int? Function(int?) getRemainingDistance;
   final VoidCallback? onTap;
   final VoidCallback? onEdit;
   final VoidCallback? onDelete;
